Move plugins.js into common
